6/3 (A lot of MTI prep)
Here are the big points:

Created Pure Pursuit algorithm that uses the same concept for a speed controller as Gluten Free RR2.
Created a bunch of actions for auto.
Created a Red auto for duck and cycling.
Created a Pipeline for detecting the duck.

Added a new subsystem for cap arm.
Added new code for the new intake.
Implemented ultrasonic sensor localization.
Added support for two cameras.

Reworked TeleOp and TeleOp controls.

Added servo position prediction and the functionality to set a servo's position over a certain amount of time to CCServo. CCServo is intended to replace the Servo class when initializing a servo similar to how the custom motor class replaces FIRST's ExMotor.